finish renaming cogs
