+++ cogs/alerts.py
@@ -7,7 +7,7 @@ from discord.commands import option
 from discord.commands.context import ApplicationContext
 from discord.ext import commands
 
-from util.embedbuilder import EmbedBuilder
+from util.embed_builder import EmbedBuilder
 from util.Logging import Log, limit
