+++ cogs/deepl.py
@@ -6,7 +6,7 @@ import thefuzz.process
 from discord import option
 from discord.ext import commands
 
-from util.embedbuilder import EmbedBuilder
+from util.embed_builder import EmbedBuilder
 from util.Logging import Log, limit
 
 LANGUAGES = [
