+++ cogs/pubchem.py
@@ -4,7 +4,7 @@ import pubchempy as pcp
 from discord import option
 from discord.ext import commands
 
-from util.embedbuilder import EmbedBuilder
+from util.embed_builder import EmbedBuilder
 from util.Logging import Log, limit
