+++ cogs/rules.py
@@ -2,7 +2,7 @@ import discord
 from discord.commands import option
 from discord.ext import commands
 
-from util.embedbuilder import EmbedBuilder
+from util.embed_builder import EmbedBuilder
 from util.Logging import Log, limit
 
 rules = {
