+++ cogs/staffapps_frontend.py
@@ -9,7 +9,7 @@ import discord.ui
 from discord.ext import commands
 from humanize import precisedelta
 
-from util.embedbuilder import EmbedBuilder
+from util.embed_builder import EmbedBuilder
 from util.Logging import Log, limit
