+++ cogs/surveys.py
@@ -3,7 +3,7 @@ from contextlib import suppress
 import discord
 from discord.ext import commands
 
-from util.embedbuilder import EmbedBuilder
+from util.embed_builder import EmbedBuilder
 from util.Logging import Log
