+++ cogs/tips.py
@@ -2,7 +2,7 @@ from discord import Member, option
 from discord.commands.context import ApplicationContext
 from discord.ext import commands
 
-from util.embedbuilder import EmbedBuilder
+from util.embed_builder import EmbedBuilder
 from util.Logging import Log, limit
 
 TIPS = {
