+++ cogs/wikipedia.py
@@ -3,7 +3,7 @@ import asyncio
 import wikipedia
 from discord.ext import commands
 
-from util.embedbuilder import EmbedBuilder
+from util.embed_builder import EmbedBuilder
 from util.Logging import Log, limit
