+++ util/Logging.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 import discord
 
-from util.embedbuilder import EmbedBuilder
+from util.embed_builder import EmbedBuilder
 
 
 def limit(_limit_level: int) -> callable:
